Update test documentation and refactor test file structure
- Improved documentation for test directory structure
- Refactored test files to match source code structure
- Updated guidelines for creating and organizing tests
