+++ tests/core/test_application.py
