+++ tests/services/generators/test_competence_generator.py
