+++ tests/services/generators/test_education_generator.py
