+++ tests/services/generators/test_skills_generator.py
