+++ tests/services/generators/test_summary_generator.py
